refactor(whispering): separate pill from overlay transport
The platform seam bundled Tauri-only status synchronization with cross-platform mic-level delivery, forcing the browser through a no-op and making shared presentation types transport-owned.

Let recording-pill own the shared model, projection, rendering, actions, and web meter. Keep recording-overlay focused on the Tauri window protocol, with mic-level delivery as the only platform seam.
