+++ apps/whispering/src/lib/recording-pill/LevelMeter.svelte
