+++ apps/whispering/src/lib/recording-pill/VadIndicator.svelte
