+++ apps/whispering/src/lib/recording-pill/mic-level.browser.ts
@@ -1,0 +1,6 @@
+import { webPillLevel } from '$lib/recording-pill/web-level.svelte';
+
+/** Feed a raw RMS sample to the in-page recording pill's reactive meter. */
+export function reportRecordingMicLevel(level: number): void {
+	webPillLevel.report(level);
+}
