+++ apps/whispering/tests/dictation-projection.test.ts
@@ -1,23 +1,23 @@
+/**
+ * Dictation Pill Projection Tests
+ *
+ * Locks the pure lifecycle projection shared by the browser host and Tauri
+ * runtime owner. A live VAD meter remains primary, with speaking and prior
+ * transcription represented as independent signals.
+ *
+ * Key behaviors:
+ * - Idle lifecycle hides the pill
+ * - Manual and VAD captures project their distinct recording states
+ * - Live VAD capture keeps outcome signals subordinate to the meter
+ */
 import { describe, expect, test } from 'bun:test';
-import { projectLifecycleToStatus } from '../src/lib/recording-overlay/projection';
+import { projectLifecycleToStatus } from '../src/lib/recording-pill/projection';
 import type {
 	DictationCapture,
 	DictationFailure,
 	DictationOutcome,
 } from '../src/lib/state/dictation-lifecycle.svelte';
 
-/**
- * Locks the dictation pill's projection invariants (ADR-0039). The projection is
- * the one place capture and outcome are flattened into the serializable status
- * both pill mounts render, so a regression here silently changes desktop and web
- * at once. These cases pin the subtle rules: a live VAD meter is never replaced,
- * a previous utterance still transcribing flags the `transcribing` signal beside
- * it, and neither success nor failure flags it (success is the landing text;
- * failure goes to the notification and the recordings row).
- *
- * The lifecycle types are structural, so the inputs are plain objects (the
- * `import type` in the projection erases at runtime, leaving a pure function).
- */
 const idle = { kind: 'idle' } satisfies DictationCapture;
 const manual = {
 	kind: 'recording',
